refactor: 인터페이스를 이용한 의존성 주입 패턴을 사용 - InputView 인터페이스 생성 및 ConsoleInputView로 콘솔 입력 처리 - OutputView 인터페이스 생성 및 ConsoleOutputView로 콘솔 출력 처리 - RacingGameService 인터페이스 생성 및 BasicRacingGameService로 게임 로직 처리 - main 메서드에서 적절한 구현체를 RacingGameController에 자유롭게 주입할 수 있도록 수정
